allow dataarray inputs into pixel_overlaps [issue 7]
xr.DataArrays are now coerced into xr.Datasets at the start of both pixel_overlaps and aggregate. This brings behavior in line with the docs of pixel_overlaps, which allow DataArray inputs. Allowing xr.DataArray inputs for aggregate() will not be added to the docs (if the xr.DataArray is unnamed, then the name 'var' will be given to the variable; not ideal behavior); a warning will be thrown if an xr.DataArray is nevertheless used and is unnamed.

Additionally, tests were added to make sure that pixel_overlaps() works with an xr.DataArray input.