implement export_weightmap()
Implement a way to export and reload the `export_weightmap()` function, which allows a temporary file of the weightmap (output of `pixel_overlaps()`) to be saved - this is particularly useful when working with very large grids where the `get_pixel_overlaps()` calculation may take a particularly long time.

Two new functions created: `export_weightmap()` and `read_file()`, which reads the exported weightmap. The weightmap class is exported as a directory containing all its constituent parts (as .shps, .ncs, .csvs).

A new testing routine is added that confirms that the export - read process is commutative.

Still missing: new dependencies on `hd5` may be explicitly needed; and one on `tables`, which, though it's installed through some other dependencies, doesn't quite work unless it's explicitly labeled as past 3.7.0.