fix export to dataset issue, insert export tests
`.to_dataset()` was not working due to too many layers of lists in the `agg.agg` geodataframe. This issue has been fixed by replacing an index with `np.squeeze()` instead. The broader problem may be that there are too many unecessary layers of lists in the `agg.agg` geodataframe, which should be simplified in the next round of backend cleanup.

Furthermore, there are now tests for `.to_dataset()` and `.to_dataframe()`