Fix issue where `weights` were not correctly added to `ds`
If the grids of `weights` and `ds` were close but not equal (floating point errors), then `weights` is not correctly added to `ds`, causing `weights` to be entirely 0s.

This has been fixed by explicitly setting the weights grid to the ds grid if their grids are `np.allclose()`.

Warnings have been added if the resultant homogenized `weights` array is all nans or all 0s.

`xa.normalize()` has been updated to ensure that if all the input vector is 0, then an output vector of all nans is returned.

New tests of `xa.normalize()` have been added to ensure it is more robust in the future.

Small fixes (changes of "|" and "&" to "or" and "and" in if statements, mainly)